feat: implement role counting functionality in authentication module
Added RolesCount method in the auth use case to retrieve the count of
roles from the repository. Implemented RoleCount method in the auth
repository to interact with the database for counting roles. Enhanced
the authGrpcHandler to utilize the new RolesCount functionality.
Additionally, introduced RbacAuthorization middleware in the
middlewareHandler to manage role-based access control, improving the
overall security and authorization flow.